+++ lib/translation/sint_translation.dart
@@ -4,4 +4,5 @@ export 'src/domain/extensions/locale_extension.dart';
 export 'src/domain/extensions/trans_extension.dart';
 export 'src/domain/interfaces/translations.dart';
 export 'src/domain/models/intl_host.dart';
+export 'src/domain/models/path_translator.dart';
 export 'src/utils/translations_constants.dart';